pyproject.toml
[tool.poetry]
name = "InspectionSheet_Generator"
version = "0.1.0"
description = ""
authors = ["SeoYoungDeok <0duck_seo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "=3.11.4"
pyyaml = "^6.0"
ipykernel = "^6.23.1"
tqdm = "^4.65.0"
python-dotenv = "^1.0.0"
pyinstaller = "^6.6.0"
pywin32 = "^306"
pyqt6 = "6.4.2"
pyqt6-tools = "^6.4.2.3.3"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
isort = "^5.12.0"
pytest = "^7.3.1"
pre-commit = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
